backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
apscheduler
python-dotenv
